web-api/requirements.txt
Flask
psycopg2-binary
redis
pika
